MTEX 4.4.0

+++ geometry/@quaternion/mean.m
@@ -1,5 +1,9 @@
-function [qm,lambda, V] = mean(q,varargin)
+function [qm, lambda, V] = mean(q,varargin)
 % mean of a list of quaternions, principle axes and moments of inertia
+%
+% Syntax
+%
+%   [m, lambda, V] = mean(q,varargin)
 %
 % Input
 %  q        - list of @quaternion
@@ -8,9 +12,9 @@ function [qm,lambda, V] = mean(q,varargin)
 %  weights  - list of weights
 %
 % Output
-%  mean     - mean orientation
-%  lambda   - principle moments of inertia
-%  V        - principle axes of inertia (@orientation)
+%  m      - mean orientation
+%  lambda - principle moments of inertia
+%  V      - principle axes of inertia (@quaternion)
 %
 % See also
 % orientation/mean
@@ -21,7 +25,7 @@ l = diag(lambda);
 pos = find(max(l)==l,1);
 qm = quaternion(V(:,pos));
 
-if check_option(varargin,'robust')
+if check_option(varargin,'robust') && length(q)>4
   omega = angle(qm,q);
   id = omega < quantile(omega,0.8);
   varargin = delete_option(varargin,'robust');
